moved shared functionality into utils package
