+++ train.py
@@ -1,105 +1,24 @@
+import assessment_utils as au
 import torch
-from torch import nn
-from torch.utils.data import DataLoader
-from torchvision import datasets
-from torchvision.transforms import ToTensor
 
-# We want to use the CIFAR-10 dataset to train the data.
-training_data = datasets.CIFAR10(
-    root="data",
-    train=True,
-    download=True,
-    transform=ToTensor()
-)
+if __name__ == "__main__":
+    # Printing the shape of the dataset allows you to know the number of input features
+    for X, y in au.train_dataloader:
+        print(f"Shape of X [N, C, H, W]: {X.shape}")
+        print(f"Shape of y: {y.shape} {y.dtype}")
+        break
 
-test_data = datasets.CIFAR10(
-    root="data",
-    train=False,
-    download=True,
-    transform=ToTensor()
-)
+    # Instantiate the model
+    model = au.Cifar_NN(au.input_features).to(au.device)
+    print(model)
 
-batch_size = 64
+    optimiser = torch.optim.SGD(model.parameters(), lr=au.lr)
 
-train_dataloader = DataLoader(training_data, batch_size=batch_size)
-test_dataloader = DataLoader(test_data, batch_size=batch_size)
+    # Perform the training
+    for t in range(au.epochs):
+        print(f"Epoch {t+1}\n----------------------------")
+        au.train(au.train_dataloader, model, au.loss_fn, optimiser)
+        au.test(au.test_dataloader, model, au.loss_fn)
+    print("Done!")
 
-for X, y in test_dataloader:
-    print(f"Shape of X [N, C, H, W]: {X.shape}")
-    print(f"Shape of y: {y.shape} {y.dtype}")
-    break
-
-device = "cpu"
-
-
-class cifar_NN(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.flatten = nn.Flatten()
-        self.linear_relu_stack = nn.Sequential(
-            nn.Linear(3*32*32, 512),
-            nn.ReLU(),
-            nn.Linear(512, 512),
-            nn.ReLU(),
-            nn.Linear(512, 10)
-        )
-
-    def forward(self, x):
-        x = self.flatten(x)
-        logits = self.linear_relu_stack(x)
-        return logits
-
-
-model = cifar_NN().to(device)
-print(model)
-
-loss_fn = nn.CrossEntropyLoss()
-optimiser = torch.optim.SGD(model.parameters(), lr=1e-2)
-
-
-def train(dataloader, model, loss_fn, optimiser):
-    size = len(dataloader.dataset)
-    model.train()
-    for batch, (X, y) in enumerate(dataloader):
-        X, y = X.to(device), y.to(device)
-
-        # Compute prediction error
-        pred = model(X)
-        loss = loss_fn(pred, y)
-
-        # Backpropagation
-        loss.backward()
-        optimiser.step()
-        optimiser.zero_grad()
-
-        if batch % 100 == 0:
-            loss, current = loss.item(), (batch + 1) * len(X)
-            print(f"Loss: {loss:7f} [{current:>5d}/{size:>5d}]")
-
-
-def test(dataloader, model, loss_fn):
-    size = len(dataloader.dataset)
-    num_batches = len(dataloader)
-    model.eval()
-    test_loss, correct = 0, 0
-    with torch.no_grad():
-        for X, y in dataloader:
-            X, y = X.to(device), y.to(device)
-            pred = model(X)
-            test_loss += loss_fn(pred, y).item()
-            correct += (pred.argmax(1) == y).type(torch.float).sum().item()
-    test_loss /= num_batches
-    correct /= size
-    print(
-        f"Test Error: \n Accuracy: {(100*correct):>0.1f}%, Avg loss:{test_loss:>8f} \n")
-
-
-epochs = 5
-for t in range(epochs):
-    print(f"Epoch {t+1}\n----------------------------")
-    train(train_dataloader, model, loss_fn, optimiser)
-    test(test_dataloader, model, loss_fn)
-print("Done!")
-
-torch.save(model.state_dict(), "model.pth")
-print("Saved PyTorch Model State to model.pth")
+    au.save_model_weights(model, au.model_save_path)
